Update documentation for the `pagination/1` component
Add a code example of how to use the component. Mark the `:params` and
`:total_count` attributes as required so the developer gets compile-time
hints that these attributes should be set in the template when using the
component. Move the LiveComponent to its own module so its functions do
not get imported for use in the application's code.

Resolves #15